Before:
``` r
library(marginaleffects)

tmp <- list()
for (i in 1:30000) {
  tmp[[i]] <- mtcars
}
tmp2 <- data.table::rbindlist(tmp)

mod <- lm(mpg ~ hp + disp + cyl, data = tmp2)

results <- bench::mark(
  out = comparisons(mod, variables = list(hp = c(100, 120), cyl = c(4, 6))),
  iterations = 10
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
results[, c(1, 3, 5)]
#> # A tibble: 1 × 3
#>   expression   median mem_alloc
#>   <bch:expr> <bch:tm> <bch:byt>
#> 1 out           9.68s    7.29GB
```
After:
``` r
results <- bench::mark(
  out = comparisons(mod, variables = list(hp = c(100, 120), cyl = c(4, 6))),
  iterations = 10
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
results[, c(1, 3, 5)]
#> # A tibble: 1 × 3
#>   expression   median mem_alloc
#>   <bch:expr> <bch:tm> <bch:byt>
#> 1 out           7.33s    7.15GB
```

Moderate gain here but I think these `unlist()` lines are called a lot in other functions.

---

FYI I discovered recently that there's a large gain with `use.names = FALSE` if you don't need the names:
``` r
test <- list(
  hp = rep("hp", 1e6),
  cyl = rep("cyl", 1e6),
  drat = rep("drat", 1e6)
)

bench::mark(
  unlist(test),
  unlist(test, use.names = FALSE),
  check = FALSE
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 2 × 6
#>   expression                           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 unlist(test)                        1.4s     1.4s     0.712    76.8MB     1.42
#> 2 unlist(test, use.names = FALSE)   19.6ms   24.1ms    10.3      22.9MB     4.43
```