@vincentarelbundock 2023-02-06
that's a cool trick! I didn't know it.